# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.DraftGroupDeleteRequest:
    description: ID группы для удаления из черновика заявки
    properties:
      group_id:
        example: 1
        type: integer
    required:
    - group_id
    type: object
  handler.DraftGroupResponse:
    description: ID заявки-черновика после операции
    properties:
      calculation_id:
        example: 1
        type: integer
    type: object
  handler.DraftGroupUpdateRequest:
    description: Данные для обновления группы в черновике (индекс CAVI)
    properties:
      cavi_index:
        example: 7.5
        type: number
      group_id:
        example: 1
        type: integer
    required:
    - group_id
    type: object
  handler.ErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      message:
        example: invalid input
        type: string
      status:
        example: fail
        type: string
    type: object
...
  handler.SuccessResponse:
    description: Стандартный успешный ответ
    properties:
      status:
        example: ok
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cavi.local
    name: API Support
  description: |-
    API для расчёта индекса CAVI (Cardio-Ankle Vascular Index).
    Система позволяет управлять группами пациентов, создавать заявки на расчёт,
    добавлять группы в заявки и проводить модерацию.
  title: CAVI Calculator API
  version: "1.0"
paths:
...
  /cavi-calculations/draft/groups:
    delete:
      consumes:
      - application/json
      description: Удаляет группу из заявки-черновика текущего пользователя.
      parameters:
      - description: ID группы
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.DraftGroupDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Группа удалена
          schema:
            $ref: '#/definitions/handler.DraftGroupResponse'
        "400":
          description: Неверный ID или группа не найдена в черновике
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить группу из черновика
      tags:
      - calculation-groups
    put:
      consumes:
      - application/json
      description: Обновляет индекс CAVI для группы в заявке-черновике текущего пользователя.
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.DraftGroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Группа обновлена
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: Неверные данные или черновик не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить группу в черновике
      tags:
      - calculation-groups
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"